Add animation to iphone image.
