--- src/lib/scoll-animation.js
@@ -1,13 +1,0 @@
-import gsap from "gsap";
-
-export const scrollAnimation = (position, target, onUpdate) => {
-  const tl = gsap.timeline()
-
-  tl.to(position, {
-    scrollTrigger: {
-      trigger: '.sound-section',
-      start: 'top bottom',
-      end: 'top'
-    }
-  })
-}
